+++ src/Core/Storefront/Checkout/Controller/CheckoutController.php
@@ -8,29 +8,39 @@ use Shopware\Core\{
 	Checkout\Cart\Exception\OrderNotFoundException,
 	Checkout\Cart\LineItemFactoryRegistry,
 	Checkout\Cart\SalesChannel\CartService,
+	Checkout\Order\Aggregate\OrderLineItem\OrderLineItemCollection,
+	Checkout\Order\Aggregate\OrderLineItem\OrderLineItemEntity,
 	Checkout\Order\OrderEntity,
 	Framework\Context,
 	Framework\DataAbstractionLayer\Search\Criteria,
 	Framework\Routing\Annotation\RouteScope,
 	Framework\Routing\Exception\MissingRequestParameterException,
-	System\SalesChannel\SalesChannelContext};
+	Framework\Validation\DataBag\RequestDataBag,
+	System\SalesChannel\SalesChannelContext
+};
 use Shopware\Storefront\{
 	Controller\StorefrontController,
 	Page\Checkout\Finish\CheckoutFinishPage,
-	Page\GenericPageLoader,};
+	Page\GenericPageLoaderInterface
+};
 use Symfony\Component\{
 	HttpFoundation\Request,
 	HttpFoundation\Response,
-	Routing\Annotation\Route};
+	Routing\Annotation\Route
+};
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use PostFinanceCheckout\Sdk\{
 	Model\Transaction,
-	Model\TransactionState};
+	Model\TransactionState
+};
 use PostFinanceCheckoutPayment\Core\{
 	Api\Transaction\Service\TransactionService,
 	Settings\Options\Integration,
 	Settings\Service\SettingsService,
-	Storefront\Checkout\Struct\CheckoutPageData};
+	Storefront\Checkout\Struct\CheckoutPageData,
+	Util\Payload\CustomProducts\CustomProductsLineItemTypes
+};
+
 
 /**
  * Class CheckoutController
@@ -83,14 +93,14 @@ class CheckoutController extends StorefrontController {
 	 * @param \Shopware\Core\Checkout\Cart\SalesChannel\CartService                         $cartService
 	 * @param \PostFinanceCheckoutPayment\Core\Settings\Service\SettingsService           $settingsService
 	 * @param \PostFinanceCheckoutPayment\Core\Api\Transaction\Service\TransactionService $transactionService
-	 * @param \Shopware\Storefront\Page\GenericPageLoader                                   $genericLoader
+	 * @param \Shopware\Storefront\Page\GenericPageLoaderInterface                          $genericLoader
 	 */
 	public function __construct(
 		LineItemFactoryRegistry $lineItemFactoryRegistry,
 		CartService $cartService,
 		SettingsService $settingsService,
 		TransactionService $transactionService,
-		GenericPageLoader $genericLoader
+		GenericPageLoaderInterface $genericLoader
 	)
 	{
 		$this->cartService             = $cartService;
@@ -321,22 +331,36 @@ class CheckoutController extends StorefrontController {
 		try {
 			// Configuration
 			$this->settings = $this->settingsService->getSettings($salesChannelContext->getSalesChannel()->getId());
-
-			$orderEntity = $this->getOrder($orderId, $salesChannelContext->getContext());
+			$orderEntity    = $this->getOrder($orderId, $salesChannelContext->getContext());
 
 			$transaction = $this->getTransaction($orderId, $salesChannelContext->getContext());
 			if (!empty($transaction->getUserFailureMessage())) {
 				$this->addFlash('danger', $transaction->getUserFailureMessage());
 			}
-			foreach ($orderEntity->getLineItems() as $orderLineItemEntity) {
+
+			$orderItems        = $orderEntity->getLineItems();
+			$hasCustomProducts = $this->hasCustomProducts($orderItems);
+
+			if ($hasCustomProducts === true) {
+				$cart = $this->addCustomProducts($orderItems, $request, $salesChannelContext);
+			}
+
+			foreach ($orderItems as $orderLineItemEntity) {
+				$type = $orderLineItemEntity->getType();
+
+				if ($type !== CustomProductsLineItemTypes::LINE_ITEM_TYPE_PRODUCT || $orderLineItemEntity->getParentid() !== null) {
+					continue;
+				}
 
 				$lineItem = $this->lineItemFactoryRegistry->create([
 					'id'           => $orderLineItemEntity->getId(),
 					'quantity'     => $orderLineItemEntity->getQuantity(),
 					'referencedId' => $orderLineItemEntity->getReferencedId(),
-					'type'         => $orderLineItemEntity->getType(),
+					'type'         => $type,
 				], $salesChannelContext);
-				$cart     = $this->cartService->add($cart, $lineItem, $salesChannelContext);
+
+				$cart = $this->cartService->add($cart, $lineItem, $salesChannelContext);
+
 			}
 
 		} catch (\Exception $exception) {
@@ -346,5 +370,146 @@ class CheckoutController extends StorefrontController {
 		}
 
 		return $this->redirectToRoute('frontend.checkout.confirm.page');
+	}
+
+	/**
+	 * @param OrderLineItemCollection $orderItems
+	 *
+	 * @return bool
+	 */
+	private function hasCustomProducts(OrderLineItemCollection $orderItems): bool
+	{
+		foreach ($orderItems as $orderItem) {
+			if ($orderItem->getType() === CustomProductsLineItemTypes::LINE_ITEM_TYPE_CUSTOMIZED_PRODUCTS) {
+				return true;
+			}
+		}
+
+		return false;
+	}
+
+	/**
+	 * @param OrderLineItemCollection $orderItems
+	 * @param string                  $parentId
+	 *
+	 * @return OrderLineItemEntity|null
+	 */
+	private function getCustomProduct(OrderLineItemCollection $orderItems, string $parentId): ?OrderLineItemEntity
+	{
+		foreach ($orderItems as $orderItem) {
+			if ($orderItem->getType() === CustomProductsLineItemTypes::LINE_ITEM_TYPE_PRODUCT && $orderItem->getParentId() === $parentId) {
+				return $orderItem;
+			}
+		}
+		return null;
+	}
+
+	/**
+	 * @param OrderLineItemCollection $orderItems
+	 * @param string                  $parentId
+	 *
+	 * @return array
+	 */
+	private function getCustomProductOptions(OrderLineItemCollection $orderItems, string $parentId): array
+	{
+		$options = [];
+		foreach ($orderItems as $orderItem) {
+			if ($orderItem->getType() === CustomProductsLineItemTypes::LINE_ITEM_TYPE_CUSTOMIZED_PRODUCTS_OPTION && $orderItem->getParentId() === $parentId) {
+				$options[] = $orderItem;
+			}
+		}
+		return $options;
+	}
+
+	/**
+	 * @param $orderItems
+	 * @param $request
+	 * @param $salesChannelContext
+	 *
+	 * @return Cart
+	 */
+	private function addCustomProducts(OrderLineItemCollection $orderItems, Request $request, SalesChannelContext $salesChannelContext): Cart
+	{
+
+		$cart = $this->cartService->getCart($salesChannelContext->getToken(), $salesChannelContext);
+		if (!\class_exists('Swag\\CustomizedProducts\\Core\\Checkout\\Cart\\Route\\AddCustomizedProductsToCartRoute')) {
+			return $cart;
+		}
+
+		$customProductsService = $this->get('Swag\CustomizedProducts\Core\Checkout\Cart\Route\AddCustomizedProductsToCartRoute');
+
+		foreach ($orderItems as $orderItem) {
+			if ($orderItem->getType() !== CustomProductsLineItemTypes::LINE_ITEM_TYPE_CUSTOMIZED_PRODUCTS) {
+				continue;
+			}
+
+			$product        = $this->getCustomProduct($orderItems, $orderItem->getId());
+			$productOptions = $this->getCustomProductOptions($orderItems, $orderItem->getId());
+			$optionValues   = $this->getOptionValues($productOptions);
+
+			$params = new RequestDataBag([]);
+
+			if (!empty($optionValues)) {
+				$params = new RequestDataBag([
+					'customized-products-template' => new RequestDataBag([
+						'id'      => $orderItem->getReferencedId(),
+						'options' => new RequestDataBag($optionValues),
+					]),
+				]);
+			}
+
+			$request->request = new RequestDataBag([
+				'lineItems' => [
+					'quantity'     => $orderItem->getQuantity(),
+					'id'           => $product->getProductId(),
+					'type'         => CustomProductsLineItemTypes::LINE_ITEM_TYPE_PRODUCT,
+					'referencedId' => $product->getReferencedId(),
+					'stackable'    => $orderItem->getStackable(),
+					'removable'    => $orderItem->getRemovable(),
+				],
+			]);
+
+			$customProductsService->add($params, $request, $salesChannelContext, $cart);
+			$cart = $this->cartService->getCart($salesChannelContext->getToken(), $salesChannelContext);
+		}
+
+		return $cart;
+	}
+
+	/**
+	 * @param array $productOptions
+	 *
+	 * @return array
+	 */
+	private function getOptionValues(array $productOptions): array
+	{
+		$optionValues = [];
+		foreach ($productOptions as $productOption) {
+			$optionType = $productOption->getPayload()['type'] ?: '';
+
+			switch ($optionType) {
+				case CustomProductsLineItemTypes::PRODUCT_OPTION_TYPE_IMAGE_UPLOAD:
+				case CustomProductsLineItemTypes::PRODUCT_OPTION_TYPE_FILE_UPLOAD:
+					$media = $productOption->getPayload()['media'] ?: [];
+					foreach ($media as $mediaItem) {
+						$optionValues[$productOption->getReferencedId()] = new RequestDataBag([
+							'media' => new RequestDataBag([
+								$mediaItem['filename'] => new RequestDataBag([
+									'id'       => $mediaItem['mediaId'],
+									'filename' => $mediaItem['filename'],
+								]),
+							]),
+						]);
+					}
+					break;
+
+				default:
+					$optionValues[$productOption->getReferencedId()] = new RequestDataBag([
+						'value' => $productOption->getPayload()['value'] ?: '',
+					]);
+			}
+		}
+
+		return $optionValues;
 	}
 }