+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 3.1.0
+- Support for Custom Products plugin
+
+# 3.0.0
+- Fix transaction versioning
+- Update SDK
+
 # 2.1.1
 - Round amounts
 - Redirect if the cart can not be recreated